rename all multiselects to Legacy
