+++ frontend/src/lib/components/apps/components/buttons/AppForm.svelte
@@ -101,7 +101,6 @@
 		componentStyle={$app.css?.formcomponent}
 	/>
 {/each}
-
 <RunnableWrapper
 	{recomputeIds}
 	{render}
